+++ src/@lib/hooks/useRef.ts
@@ -1,8 +1,6 @@
-// import { useState } from "react";
-
 import { useState } from "react";
 
 export function useRef<T>(initialValue: T): { current: T } {
-  const [ ref ] = useState({ current: initialValue });
+  const [ref] = useState({ current: initialValue });
   return ref;
 }
