+++ src/contexts/UserContext.tsx
@@ -1,0 +1,36 @@
+import { createContext, useContext, useState } from "react";
+import { useCallback } from "../@lib";
+import { useNotification } from "./NotificationContext";
+import type { User, UserContextType } from "../types";
+
+const UserContext = createContext<UserContextType | null>(null);
+
+export const useUser = () => {
+  const ctx = useContext(UserContext);
+  if (!ctx) throw new Error("useUser must be used within UserProvider");
+  return ctx;
+};
+
+export const UserProvider: React.FC<{ children: React.ReactNode }> = ({
+  children
+}) => {
+  const [ userState, setUser ] = useState<User | null>(null);
+  const { addNotification } = useNotification();
+
+  const login:(message: string) => void = useCallback((email: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+    addNotification("성공적으로 로그인되었습니다", "success");
+  }, [setUser, addNotification]);
+
+
+  const logout:() => void = useCallback(() => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  }, [setUser, addNotification]);
+
+  return (
+    <UserContext.Provider value={{ user: userState, login, logout }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
